refactor: split more interfaces into their own file
