+++ src/interfaces/decorator-options/exclude-option.interface.ts
@@ -1,0 +1,14 @@
+/**
+ * Possible transformation options for the @Exclude decorator.
+ */
+export interface ExcludeOptions {
+  /**
+   * Expose this property only when transforming from plain to class instance.
+   */
+  toClassOnly?: boolean;
+
+  /**
+   * Expose this property only when transforming from class instance to plain object.
+   */
+  toPlainOnly?: boolean;
+}
