+++ src/interfaces/index.ts
@@ -1,3 +1,9 @@
+export * from './decorator-options/expose-option.interface';
+export * from './decorator-options/exclude-option.interface';
+export * from './decorator-options/transform-option.interface';
+export * from './decorator-options/type-discriminator-descriptor.interface';
+export * from './decorator-options/type-option.interface';
 export * from './class-constructor.type';
 export * from './class-transformer-options.interface';
 export * from './target-map.interface';
+export * from './type-help-options.interface';
