+++ src/interfaces/type-help-options.interface.ts
@@ -1,0 +1,6 @@
+// TODO: Document this interface. What does each property means?
+export interface TypeHelpOptions {
+  newObject: any;
+  object: Record<string, any>;
+  property: string;
+}
