+++ src/metadata/ExcludeMetadata.ts
@@ -1,4 +1,4 @@
-import { ExcludeOptions } from './ExposeExcludeOptions';
+import { ExcludeOptions } from '../interfaces';
 
 export class ExcludeMetadata {
   constructor(public target: Function, public propertyName: string, public options: ExcludeOptions) {}
