+++ src/metadata/ExposeMetadata.ts
@@ -1,4 +1,4 @@
-import { ExposeOptions } from './ExposeExcludeOptions';
+import { ExposeOptions } from '../interfaces';
 
 export class ExposeMetadata {
   constructor(public target: Function, public propertyName: string, public options: ExposeOptions) {}
